migratio

+++ backend_web/db/migrations/20220910010203_add_raffle_configuration.php
@@ -23,7 +23,7 @@ final class AddRaffleConfiguration extends AbsMigration
     private function _add_columns_to_app_promotion_cap(): void
     {
         $table = $this->table("app_promotion_cap");
-        $table->addColumn("id_raffle", "tinyint", [
+        $table->addColumn("id_raffle", "integer", [
             "limit" => 3,
             "null" => true,
             "default" => null,
